--- a/doc/tools/pyrmvl_fns.cfg
+++ b/doc/tools/pyrmvl_fns.cfg
@@ -48,8 +48,8 @@ rm                     region          func,x0[,delta]         a,b
 rm                     fminbnd         func,x1,x2[,options]    x,fval
 rm                     fminunc         func,x0[,options]       x,fval
 rm                     fmincon         func,x0,c,ceq[,options] x,fval
-rm                     lsqnonlin       funcs,x0[,options]      x
-rm                     lsqnonlinRKF    funcs,x0,rb[,options]   x
+rm                     lsqnonlin       func,x0[,options]       x
+rm                     lsqnonlinRKF    func,x0,rb[,options]    x
 rm                     dft             xt                      Xf
 rm                     idft            Xf                      xt
 rm                     Gx              x,type                  G