### Pull Request 合并请求准备清单

详情参见[此处](https://github.com/cv-rmvl/rmvl/wiki/How_to_contribute#3-making-a-good-pull-request)

- [x] 我同意在 Apache 2 开源许可下为本项目做贡献
- [x] 此 pull request 是在正确的分支上提出的
- [x] 此 pull request 有对应的错误报告或其他待改进的内容
- [x] 我本地的 RMVL 进行了单元测试、性能测试，有对应的测试数据
- [x] 我提交的 feature 有很好的文档记录，并且可以使用 CMake 项目构建示例代码

### 具体内容

Finish and close #213 

### 基准测试

`rm::FuncNds` 在新旧表示方法上效果相近，以 `lsqnonlin (sine)` 为例

```
-----------------------------------------------------------------------------------------------------
Benchmark                                                           Time             CPU   Iterations
-----------------------------------------------------------------------------------------------------
old lsqnonlin (sine) - by rmvl/iterations:50                    79909 ns        76771 ns           50
new lsqnonlin (sine) - by rmvl/iterations:50                    74344 ns        74348 ns           50
```